web: Rename font_renderer.rs to canvas_font_renderer.rs
This file name should be clearer.
